requirements.txt
robotframework>=6.0.0
robotframework-seleniumlibrary>=6.0.0
robotframework-whitelibrary>=2.1.0
robotframework-pabot>=2.12.0
selenium>=4.9.0
webdrivermanager>=0.10.0
pyyaml>=6.0
robotframework-faker>=5.0.0
robotframework-metrics>=3.3.3
python-dotenv>=1.0.0
pywinauto>=0.6.8
pytest>=7.3.1 